+++ src/main/java/org/folio/linked/data/controller/BatchJobController.java
@@ -4,9 +4,9 @@ import static java.lang.String.valueOf;
 import static org.folio.linked.data.util.Constants.STANDALONE_PROFILE;
 
 import lombok.RequiredArgsConstructor;
-import org.folio.linked.data.domain.dto.ReindexJobStatusDto;
-import org.folio.linked.data.rest.resource.ReindexApi;
-import org.folio.linked.data.service.batch.ReindexJobService;
+import org.folio.linked.data.domain.dto.BatchJobStatusDto;
+import org.folio.linked.data.rest.resource.BatchJobApi;
+import org.folio.linked.data.service.batch.BatchJobService;
 import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -16,24 +16,30 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @Profile("!" + STANDALONE_PROFILE)
-public class ReindexingController implements ReindexApi {
+public class BatchJobController implements BatchJobApi {
 
-  private final ReindexJobService reindexJobService;
+  private final BatchJobService batchJobService;
 
   @Override
   public ResponseEntity<String> fullReindex(String resourceType) {
-    var jobExecutionId = valueOf(reindexJobService.start(true, resourceType));
+    var jobExecutionId = valueOf(batchJobService.startReindex(true, resourceType));
     return ResponseEntity.ok(jobExecutionId);
   }
 
   @Override
   public ResponseEntity<String> incrementalReindex(String resourceType) {
-    var jobExecutionId = valueOf(reindexJobService.start(false, resourceType));
+    var jobExecutionId = valueOf(batchJobService.startReindex(false, resourceType));
     return ResponseEntity.ok(jobExecutionId);
   }
 
   @Override
-  public ResponseEntity<ReindexJobStatusDto> getReindexJobStatus(Long jobExecutionId) {
-    return ResponseEntity.ok(reindexJobService.getStatus(jobExecutionId));
+  public ResponseEntity<String> graphCleaning() {
+    var jobExecutionId = valueOf(batchJobService.startGraphCleaning(1));
+    return ResponseEntity.ok(jobExecutionId);
+  }
+
+  @Override
+  public ResponseEntity<BatchJobStatusDto> getBatchJobStatus(Long jobExecutionId) {
+    return ResponseEntity.ok(batchJobService.getStatus(jobExecutionId));
   }
 }