+++ src/main/java/org/folio/linked/data/service/batch/BatchJobService.java
@@ -1,0 +1,12 @@
+package org.folio.linked.data.service.batch;
+
+import org.folio.linked.data.domain.dto.BatchJobStatusDto;
+
+public interface BatchJobService {
+
+  Long startReindex(boolean isFullReindex, String resourceType);
+
+  Long startGraphCleaning(int executionRound);
+
+  BatchJobStatusDto getStatus(Long jobExecutionId);
+}