+++ NEWS.md
@@ -107,6 +107,7 @@
 - Retrieve FOLIO base-url from `GET /base-url` API [MODLD-804](https://folio-org.atlassian.net/browse/MODLD-804)
 - Use LabelGeneratorService to compute the label of WORK, INSTANCE and TITLE resources [MODLD-998](https://folio-org.atlassian.net/browse/MODLD-998)
 - Update DissertationMapperUnit based on the updated mapping rules [MODLD-1014](https://folio-org.atlassian.net/browse/MODLD-1014)
+- Graph cleaning spring batch job and schedule [MODLD-1013](https://folio-org.atlassian.net/browse/MODLD-1013)
 - Add import from URL API [MODLD-1019](https://folio-org.atlassian.net/browse/MODLD-1019)
 
 ## 1.0.4 (04-24-2025)